add route for getting the list of expenses
